rag_pipeline/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
groq>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.0.0          # terminal Markdown rendering (tables, headers, bold)
pytest>=8.0.0         # test runner
